Pss支持websocket协议达成
Pss支持websocket协议，新的PacketParse在OtherPacketParse目录下，并且测试插件websocket 在
example目录下

+++ purenessscopeserver/purenessscopeserver/PurenessScopeServer/Common/MainConfig.cpp
@@ -431,7 +431,7 @@ bool CMainConfig::Init(const char* szConfigPath)
 	pData = m_MainConfig.GetData("ClientInfo", "MaxBuffRecv");
 	if(pData != NULL)
 	{
-		m_u4ServerRecvBuff = (uint8)ACE_OS::atoi(pData);
+		m_u4ServerRecvBuff = (uint32)ACE_OS::atoi(pData);
 	}
 
 	//接收客户端信息相关配置
